+++ src/redturtle/prenotazioni/events/events_logger.py
@@ -3,7 +3,6 @@ from json import dumps
 from time import time
 
 from plone import api
-from redturtle.prenotazioni import logger
 
 
 def log_data_for_booking(obj, data):
@@ -48,7 +47,7 @@ def on_modify(obj, event):
     This handler logs a cvs string for
     every IPrenotazione document modified
     """
-    old_version = getattr(obj, "version_id", 0) - 1
+    old_version = getattr(obj.aq_base, "version_id", 0) - 1
     if old_version < 0:
         return
 
@@ -64,12 +63,9 @@ def on_modify(obj, event):
         "title",
     ]
     pr = api.portal.get_tool(name="portal_repository")
-    try:
-        old = pr.retrieve(obj, old_version).object
-    except Exception:
-        # ArchivistRetrieveError see https://github.com/RedTurtle/redturtle.prenotazioni/pull/178
-        logger.exception("error on_modify %s", obj.absolute_url())
-        return
+
+    old = pr.retrieve(obj, old_version).object
+
     changes = []
     for fname in fnames:
         c_value = obj.getField(fname, obj).get(obj)